requirements.txt
apache-airflow==2.10.0
pandas==2.0.3
